Update utils.py
Updated Import from erpnext.utilities.product at line 3 to 
update to line 4 from webshop.webshop.utils import adjust_qty_for_expired_items.
this is as a result of removing ecommerce module in favor of webshop in erpnext v15. the change is necessary as it crashes the server. if the summit-api is installed.